Refactor hardware breakpoint manager functions
Move non-architecture-specific functions into a common file
