+++ Sources/Core/X86/HardwareBreakpointManager.cpp
@@ -8,7 +8,7 @@
 // PATENTS file in the same directory.
 //
 
-#include "DebugServer2/Architecture/X86/HardwareBreakpointManager.h"
+#include "DebugServer2/Core/HardwareBreakpointManager.h"
 #include "DebugServer2/Target/Process.h"
 #include "DebugServer2/Target/Thread.h"
 #include "DebugServer2/Utils/Bits.h"
@@ -22,42 +22,13 @@ using ds2::Utils::EnableBit;
 using ds2::Utils::DisableBit;
 
 namespace ds2 {
-namespace Architecture {
-namespace X86 {
 
-static const int kMaxHWStoppoints = 4; // dr0, dr1, dr2, dr3
 static const int kStatusRegIdx = 6;
 static const int kCtrlRegIdx = 7;
 
-HardwareBreakpointManager::HardwareBreakpointManager(
-    Target::ProcessBase *process)
-    : super(process), _locations(kMaxHWStoppoints, 0) {}
-
-HardwareBreakpointManager::~HardwareBreakpointManager() {}
-
-ErrorCode HardwareBreakpointManager::add(Address const &address, Type type,
-                                         size_t size, Mode mode) {
-  DS2ASSERT(_sites.size() <= kMaxHWStoppoints);
-
-  if (mode == kModeRead) {
-    DS2LOG(Warning,
-           "read-only watchpoints are unsupported, setting as read-write");
-    mode = static_cast<Mode>(mode | kModeWrite);
-  }
-
-  return super::add(address, type, size, mode);
+size_t HardwareBreakpointManager::maxWatchpoints() {
+  return 4; // dr0, dr1, dr2, dr3
 }
-
-ErrorCode HardwareBreakpointManager::remove(Address const &address) {
-  auto loc = std::find(_locations.begin(), _locations.end(), address);
-  if (loc != _locations.end()) {
-    _locations[loc - _locations.begin()] = 0;
-  }
-
-  return super::remove(address);
-}
-
-int HardwareBreakpointManager::maxWatchpoints() { return kMaxHWStoppoints; }
 
 ErrorCode HardwareBreakpointManager::enableLocation(Site const &site, int idx,
                                                     Target::Thread *thread) {
@@ -91,37 +62,6 @@ ErrorCode HardwareBreakpointManager::enableLocation(Site const &site, int idx,
   return kSuccess;
 }
 
-ErrorCode HardwareBreakpointManager::enableLocation(Site const &site) {
-  int idx;
-
-  auto loc = std::find(_locations.begin(), _locations.end(), site.address);
-  if (loc == _locations.end()) {
-    idx = getAvailableLocation();
-    if (idx < 0) {
-      return kErrorInvalidArgument;
-    }
-  } else {
-    idx = loc - _locations.begin();
-  }
-
-  std::set<Target::Thread *> threads;
-  _process->enumerateThreads([&](Target::Thread *thread) {
-    if (thread->state() == Target::Thread::kStopped) {
-      threads.insert(thread);
-    }
-  });
-
-  for (auto thread : threads) {
-    ErrorCode error = enableLocation(site, idx, thread);
-    if (error != kSuccess) {
-      return error;
-    }
-  }
-
-  _locations[idx] = site.address;
-  return kSuccess;
-}
-
 ErrorCode HardwareBreakpointManager::disableLocation(int idx,
                                                      Target::Thread *thread) {
   ErrorCode error;
@@ -149,32 +89,6 @@ ErrorCode HardwareBreakpointManager::disableLocation(int idx,
   if (error != kSuccess) {
     DS2LOG(Error, "failed to write to debug control register");
     return error;
-  }
-
-  return kSuccess;
-}
-
-ErrorCode HardwareBreakpointManager::disableLocation(Site const &site) {
-  auto loc = std::find(_locations.begin(), _locations.end(), site.address);
-  if (loc == _locations.end()) {
-    return kErrorInvalidArgument;
-  }
-  int idx = loc - _locations.begin();
-
-  std::set<Target::Thread *> threads;
-  _process->enumerateThreads([&](Target::Thread *thread) {
-    if (thread->state() == Target::Thread::kStopped) {
-      threads.insert(thread);
-    }
-  });
-
-  for (auto thread : threads) {
-    DS2ASSERT(static_cast<uintptr_t>(site.address) ==
-              static_cast<uintptr_t>(thread->readDebugReg(idx)));
-    ErrorCode error = disableLocation(idx, thread);
-    if (error != kSuccess) {
-      return error;
-    }
   }
 
   return kSuccess;
@@ -248,18 +162,6 @@ ErrorCode HardwareBreakpointManager::disableDebugCtrlReg(uint32_t &ctrlReg,
   return kSuccess;
 }
 
-int HardwareBreakpointManager::getAvailableLocation() {
-  DS2ASSERT(_locations.size() == kMaxHWStoppoints);
-  if (_sites.size() == kMaxHWStoppoints) {
-    return -1;
-  }
-
-  auto it = std::find(_locations.begin(), _locations.end(), 0);
-  DS2ASSERT(it != _locations.end());
-
-  return (it - _locations.begin());
-}
-
 int HardwareBreakpointManager::hit(Target::Thread *thread, Site &site) {
   if (_sites.size() == 0) {
     return -1;
@@ -271,7 +173,7 @@ int HardwareBreakpointManager::hit(Target::Thread *thread, Site &site) {
 
   uint32_t status_reg = thread->readDebugReg(kStatusRegIdx);
 
-  for (int i = 0; i < kMaxHWStoppoints; ++i) {
+  for (size_t i = 0; i < maxWatchpoints(); ++i) {
     if (status_reg & (1 << i)) {
       DS2ASSERT(_locations[i] != 0);
       site = _sites.find(_locations[i])->second;
@@ -307,7 +209,5 @@ ErrorCode HardwareBreakpointManager::isValid(Address const &address,
   }
 
   return super::isValid(address, size, mode);
-}
-}
 }
 }
